fix(portfolio): complete module

+++ internal/migrations/000001_create_portfolio_work_images_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE portfolio_work_images (
-    id SERIAL PRIMARY KEY,
-    portfolio_work_id INT NOT NULL REFERENCES portfolio_works(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY,
+    portfolio_work_id BIGINT NOT NULL REFERENCES portfolio_works(id) ON DELETE CASCADE,
     image_url TEXT NOT NULL,
     is_thumbnail BOOLEAN DEFAULT FALSE,
     caption TEXT,
